refactor: move sync model
